Updated neural network implementation and API logic

+++ src/Main.java
@@ -12,7 +12,7 @@ public class Main {
         api.JSONtoString(jsonStock);
 
         //generate dataset
-        int[][] dataset = data.generateDataset();
+        double[][] dataset = data.generateDataset();
 
 
         //generate input data
