+++ src/data.java
@@ -5,10 +5,10 @@ import java.util.Arrays;
 public class data {
 
     //this function will be used to generate the dataset
-    public static int[][] generateDataset() throws Exception {
+    public static double[][] generateDataset() throws Exception {
 
         //this variable is a 2d array which will be used as the dataset
-        int[][] dataset = new int[2 * (api.dataLength + 30)][6];
+        double[][] dataset = new double[2 * (api.dataLength + 30)][6];
 
         //those two for loops is used to create the dataset
         for (int i = 0; i <= api.dataLength; i++) {
@@ -25,32 +25,32 @@ public class data {
 
                     //check if k equals to 0 --> openArr
                     case 0:
-                        dataset[i][k] = api.openArr.getInt(i); //in case k == 0 s
+                        dataset[i][k] = api.openArr.getDouble(i); //in case k == 0 s
                         break; //break
 
                         //check if k equals to 1 --> highArr
                     case 1:
-                        dataset[i][k] = api.highArr.getInt(i); //in case k == 0 s
+                        dataset[i][k] = api.highArr.getDouble(i); //in case k == 0 s
                         break; //break
 
                         //check if k equals to 2 --> lowArr
                     case 2:
-                        dataset[i][k] = api.lowArr.getInt(i); //in case k == 0 s
+                        dataset[i][k] = api.lowArr.getDouble(i); //in case k == 0 s
                         break; //break
 
                         //check if k equals to 3 --> closeArr
                     case 3:
-                        dataset[i][k] = api.closeArr.getInt(i); //in case k == 0 s
+                        dataset[i][k] = api.closeArr.getDouble(i); //in case k == 0 s
                         break; //break
 
                         //check if k equals to 4 --> volumeArr
                     case 4:
-                        dataset[i][k] = api.volumeArr.getInt(i); //in case k == 0 s
+                        dataset[i][k] = api.volumeArr.getDouble(i); //in case k == 0 s
                         break; //break
 
                         //check if k equals to 5 --> timestampArr
                     case 5:
-                        dataset[i][k] = api.timestampArr.getInt(i); //in case k == 0 s
+                        dataset[i][k] = api.timestampArr.getDouble(i); //in case k == 0 s
                         break; //break
                 }
